# mnt/private/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API de développement"
        }
    ],
    "paths": {
...
        "/api/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Updates the user's information (email, phone, address, postal code)",
                "operationId": "b08980cc0e7f23574681c256ab16e35a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "inputEmail",
                                    "inputPhoneNumber",
                                    "inputAddress",
                                    "inputPostalCode"
                                ],
                                "properties": {
                                    "inputEmail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "inputPhoneNumber": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "inputAddress": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "inputPostalCode": {
                                        "type": "string",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User information updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Updates the user's password",
                "operationId": "7cf63ef77ae59c4d77e571a5d2a23993",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "inputActualPassword",
                                    "inputNewPassword",
                                    "inputPasswordVerif"
                                ],
                                "properties": {
                                    "inputActualPassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "currentpassword123"
                                    },
                                    "inputNewPassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "inputPasswordVerif": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                }
            }
        },
...
    },
...
}